requirements.txt
flask
fasttext-wheel
flask-cors
gunicorn
